Siteuser 엔티티 추가

+++ src/test/java/com/ll/exam/QuestionRepositoryTests.java
@@ -22,7 +22,7 @@ class QuestionRepositoryTests {
 
     @Autowired
     private QuestionRepository questionRepository;
-    private static int lastSampleDataId;
+    private static Long lastSampleDataId;
 
     @BeforeEach
     void beforeEach(){
@@ -30,7 +30,7 @@ class QuestionRepositoryTests {
         createSampleData(); // 데이터 만들기
     }
 
-    public static int createSampleData(QuestionRepository questionRepository) {
+    public static Long createSampleData(QuestionRepository questionRepository) {
         Question q1 = new Question();
         q1.setSubject("sbb가 무엇인가요?");
         q1.setContent("sbb에 대해서 알고 싶습니다.");
@@ -149,7 +149,7 @@ class QuestionRepositoryTests {
     void Pageable(){
 
         // 한 페이지에 몇 개의 아이템이 나와야 하는지 알려주고 현재 몇 페이지인지 알려준다
-        Pageable pageable = PageRequest.of(0, lastSampleDataId);
+        Pageable pageable = PageRequest.of(0, Math.toIntExact(lastSampleDataId));
 
         Page<Question> page = questionRepository.findAll(pageable);
